+++ src/Pages/Barcelona.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import Track from '../Components/Track'
+const trackName = "Circuit de Catalunya";
+const trackID= "barcelona";
+const URL = "/images/1000px-Circuit_Catalunya_2007.svg.png";
+
+const Barcelona = () => {
+  return (
+    <Track
+      trackName={trackName}
+      trackID={trackID}
+      URL={URL}
+    />
+  )
+}
+
+export default Barcelona
