+++ src/Pages/routes.js
@@ -15,4 +15,3 @@ export const SEBRING = '/sebring'
 export const BATHURST = '/mount-panorama'
 export const ZOLDER = '/zolder'
 export const DAYTONA = '/daytona'
-export const BARCELONA = '/circuit-de-catalunya'
