--- src/styles/_drawer.scss
@@ -1,21 +1,0 @@
-#drawer {
-
-    padding: $baseline;
-    position: fixed;
-    top: 0;
-    right: 0;
-    width: 70vw;
-    background: #fff;
-    height: 100vh;
-    box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.5);
-    transition: all 0.3 ease;
-    z-index: 10;
-
-    &.hide {
-      transform: translate3d(-100vw, 0, 0);
-    }
-
-    &.show {
-      transform: translate3d(0vw,0,0);
-    }
-}
