--- src/styles/_notes.scss
@@ -1,54 +1,0 @@
-// .turn-number {
-//   font-size: 2rem;
-//   font-weight: bold;
-//   min-width: 50px;
-// }
-//
-// .track-logs {
-//   display: flex;
-//   flex-direction: row;
-//
-//   div:last-child {
-//     margin-left: auto;
-//   }
-// }
-//
-// .track-turn-notes {
-//   textarea {
-//     width: 400px;
-//     min-height: 100px;
-//     font-size: 1.6rem;
-//     border: 1px solid $colour-gray-300;
-//   }
-// }
-
-.new-observation,
-.new-corner {
-
-  label {
-    margin-bottom: $baseline/4;
-  }
-  textarea {
-    width: 100%;
-    font-family: $fontstack;
-    font-size: 1.4rem;
-    padding: $baseline/4;
-    background: #F7F7F7;
-    border-radius: 2px;
-  }
-
-  input[type="text"] {
-    background: #F7F7F7;
-    border-radius: 2px;
-  }
-
-  .form-actions {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    border-bottom: 1px solid #E9E9E9;
-    padding-bottom: $baseline;
-    margin-bottom: $baseline*2;
-  }
-}
